+++ lib/send_mail_of_qiita_minutes/email_address.rb
@@ -1,10 +1,14 @@
+require 'send_mail_of_qiita_minutes/config_base'
+
 module SendMailOfQiitaMinutes
   class EmailAddress
+    include SendMailOfQiitaMinutes::ConfigBase
 
     TARGET_NAME_EMAIL_ADDRESSES = 'email_addresses'.freeze
 
     def initialize(options:)
       @options = options
+      @target_name = TARGET_NAME_EMAIL_ADDRESSES
     end
 
     def execute
@@ -33,17 +37,7 @@ module SendMailOfQiitaMinutes
     end
 
     def self.read_config
-      result_hash = nil
-      if File.exist?(CONFIG_FILE_NAME)
-        data = File.open CONFIG_FILE_NAME do |f|
-          f.read
-        end
-
-        hash = JSON.load data
-        result_hash = hash[TARGET_NAME_EMAIL_ADDRESSES]
-      end
-
-      result_hash
+      ConfigBase.get_config(TARGET_NAME_EMAIL_ADDRESSES)
     end
 
     private
@@ -53,8 +47,5 @@ module SendMailOfQiitaMinutes
       SendMailOfQiitaMinutes.write_json_for_append target: TARGET_NAME_EMAIL_ADDRESSES, data: hash
     end
 
-    def delete_config
-      SendMailOfQiitaMinutes.delete_config target:TARGET_NAME_EMAIL_ADDRESSES
-    end
   end
 end
