+++ lib/send_mail_of_qiita_minutes/sender_email_address.rb
@@ -1,12 +1,15 @@
 require 'json'
+require 'send_mail_of_qiita_minutes/config_base'
 
 module SendMailOfQiitaMinutes
   class SenderEmailAddress
+    include SendMailOfQiitaMinutes::ConfigBase
+
     TARGET_NAME_SENDER_EMAIL_ADDRESS = 'sender_email_address'.freeze
 
     def initialize(options:)
-      p options
       @options = options
+      @target_name = TARGET_NAME_SENDER_EMAIL_ADDRESS
     end
 
     def execute
@@ -28,22 +31,8 @@ module SendMailOfQiitaMinutes
       end
     end
 
-    def self.read_config(symbolize_names: false)
-      result_hash = nil
-      if File.exist?(CONFIG_FILE_NAME)
-        data = File.open CONFIG_FILE_NAME do |f|
-          f.read
-        end
-
-        hash = JSON.parse(data)
-        result_hash = hash[TARGET_NAME_SENDER_EMAIL_ADDRESS]
-      end
-
-      if symbolize_names
-        result_hash.map{|k,v| [k.to_sym, v] }.to_h
-      else
-        result_hash
-      end
+    def self.read_config
+      ConfigBase.get_config(TARGET_NAME_SENDER_EMAIL_ADDRESS)
     end
 
     private
@@ -51,10 +40,6 @@ module SendMailOfQiitaMinutes
     def write_config(sender_email_address:)
       hash = {TARGET_NAME_SENDER_EMAIL_ADDRESS => sender_email_address}
       SendMailOfQiitaMinutes.write_json_for_append target: TARGET_NAME_SENDER_EMAIL_ADDRESS, data: hash
-    end
-
-    def delete_config
-      SendMailOfQiitaMinutes.delete_config  target:TARGET_NAME_SENDER_EMAIL_ADDRESS
     end
   end
 end
